blog

+++ app/Http/Controllers/BlogController.php
@@ -8,6 +8,8 @@ use Illuminate\Support\Facades\Validator;
 
 use App\Models\User;
 use App\Models\Post;
+use App\Models\Tag;
+use App\Models\Post_tag;
 
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
@@ -16,13 +18,18 @@ class BlogController extends Controller
 {
     public function index()
     {
-        $posts = Post::where("user_id", Auth::user()->id)
-            ->get();
-        $user = User::findOrFail( Auth::user()->id);
+        $posts = Post::where("post_status_id", '<>', 3)->get();
+        $approvePostNumber = 0;
 
-    return view("Admin.Blogs.IndexAll")
-        ->with("posts", $posts)
-        ->with("user",$user);
+        foreach ($posts as $post) {
+            if ($post->post_status_id == 1 ) {
+                $approvePostNumber = $approvePostNumber + 1;
+            }
+        }
+        
+        return view("Admin.Blogs.IndexAll")
+            ->with("posts", $posts)
+            ->with("approvePostNumber", $approvePostNumber);
     }
     
     
@@ -35,11 +42,9 @@ class BlogController extends Controller
     {
         $posts = Post::where("user_id", Auth::user()->id)
             ->get();
-    $user = User::findOrFail( Auth::user()->id);
 
     return view("Admin.Blogs.Index")
-        ->with("posts", $posts)
-        ->with("user",$user);
+        ->with("posts", $posts);
     }
 
     /**
@@ -49,7 +54,9 @@ class BlogController extends Controller
      */
     public function create()
     {
-        return view('Admin.Blogs.Create');
+        $tags = Tag::orderBy("name")->get();
+        return view('Admin.Blogs.Create')
+            ->with("tags", $tags);
     }
 
     /**
@@ -67,6 +74,7 @@ class BlogController extends Controller
             ];
 
         }else{
+            //TODO: ha túl hosszú a lead akkor is hibaüzenettel térjen vissza!
             $rules = [
                 "title" => "required|unique:posts",
                 "lead" => "required",
@@ -74,8 +82,6 @@ class BlogController extends Controller
                 "cover" => "required"
             ];
         }
-        
-
         $messages = [
             "title.required" => "A cím mező kitöltése kötelező!",
             "title.unique" => "A megadott cím már létezik, válasszon másikat!",
@@ -87,20 +93,22 @@ class BlogController extends Controller
         $validator = Validator::make($request->all(), $rules, $messages);
 
         if ( $validator->fails() )
+            //TODO: képet ne töltse fel ilyenkor.
             return back()
                 ->withErrors($validator)
                 ->withInput();
-        
+
         //kép feltöltés
         $imageName = "";
-        
         
         if($request->cover){
             $image = $request->file('cover');
             $imageInfo = pathinfo($image->getClientOriginalName());
-            $imageName = \Carbon\Carbon::now()->format("U").Str::slug($request->name).".".$imageInfo['extension'];
+            $imageName = \Carbon\Carbon::now()->format("U").Str::slug($request->cover).".".$imageInfo['extension'];
             $destinationPath = public_path('/images/posts');
             $image->move($destinationPath, $imageName);
+
+            $imageName = "/images/posts/".$imageName;
         }
         
         $lead = "";
@@ -117,15 +125,22 @@ class BlogController extends Controller
         $post->lead = $lead;
         $post->body = $body;
         $post->user_id = Auth::user()->id;
-        
-        $post->cover = "/images/posts/".$imageName;
+        $post->post_status_id = 1;
+        $post->cover = $imageName;
 
         if($request->has('draft'))
             $post->post_status_id = 3;
         
         $post->save();
+        
+        
+        //TAG-ek felvétele
+        $tags = $request->tags;
+        $post->tag()->sync($tags);
+        
         if($post->post_status_id = 3)
             return redirect()->route("blogByUser")->withSuccess("A bejegyzés mentése sikerült! Státusza 'Piszkozat'-ra módosult!");
+
         return redirect()->route("blogByUser")->withSuccess("A bejegyzés létrehozása sikerült! Státusza 'Jóváhagyásra vár'-ra módosult!");
         
     }
@@ -139,9 +154,13 @@ class BlogController extends Controller
     public function edit($id)
     {
         $post = Post::findOrFail($id);
+        $post_tags = Post_tag::where("post_id", $id)->pluck("tag_id")->toArray();
+        $tags = Tag::orderBy("name")->get();
         
         return view('Admin.Blogs.Edit')
-                    ->with('post', $post);
+                    ->with('post', $post)
+                    ->with('tags', $tags)
+                    ->with('post_tags', $post_tags);
     }
 
     /**
@@ -157,12 +176,11 @@ class BlogController extends Controller
 
         //1. validálás
         $rules = [
-            "title" => "required",
-            
+            "title" => "required|max:255",
         ];
         if($request->has('publish')){
             if ( $post->lead != $request->lead)
-                $rules["lead"] = "required";
+                $rules["lead"] = "required|max:500";
 
             if ( $post->body != $request->body)
                 $rules["body"] = "required";
@@ -174,7 +192,9 @@ class BlogController extends Controller
         $messages = [
             "title.required" => "A cím mező kitöltése kötelező!",
             "title.unique" => "A megadott cím már létezik, válasszon másikat!",
+            "title.max" => "A cím maximális hossza :max karakter lehet!",
             "lead.required" => "A bevezető mező kitöltése kötelező!",
+            "lead.max" => "A bevezető maximális hossza :max karakter lehet!",
             "body.required" => "A szövegtörzs mező nem lehet üres!",
         ];
 
@@ -188,9 +208,11 @@ class BlogController extends Controller
 
         //kép feltöltés
         if ( $request->cover ) {
-            $path = public_path()."/images/posts/".$post->cover;
-            if ( file_exists($path) ) 
-                unlink($path);
+            if ( $post->cover!="" ) {
+                $path = public_path()."/images/posts/".$post->cover;
+                if ( file_exists($path) ) 
+                    unlink($path);
+            }
 
             $image = $request->file('cover');
             $imageInfo = pathinfo($image->getClientOriginalName());
@@ -212,7 +234,6 @@ class BlogController extends Controller
         $post->slug = Str::slug($request->title);
         $post->lead = $lead;
         $post->body = $body;
-        $post->user_id = Auth::user()->id;
         
         if($request->has('draft')){
             $post->post_status_id = 3;
@@ -220,6 +241,12 @@ class BlogController extends Controller
             $post->post_status_id = 1;
         }
         $post->save();
+
+        //TAG-ek módosítása
+        $tags = $request->tags;
+        $post->tag()->sync($tags);
+
+
         if($post->post_status_id == 3)
             return redirect()->route("blogByUser")->withSuccess("A piszkozat mentése sikerült!");
         return redirect()->route("blogByUser")->withSuccess("A bejegyzés módosítása sikerült! Státusza 'Jóváhagyásra vár'-ra változott.");
@@ -262,7 +289,7 @@ class BlogController extends Controller
         $post->save();
 
         return redirect()
-            ->route("postStatus")
+            ->route("blogAll")
             ->withSuccess("A jóváhagyás sikerült!");
     }
